+++ source/extensions/lightspeed.trex.hotkeys/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.4.5]
+### Added
+- Added a global Ctrl+A event for Select All subscribers.
+
 ## [0.4.4]
 ### Changed
 - Updated extension metadata for Kit SDK 110 compatibility.